requirements.txt
# Basic dependencies for local development
torch>=2.0.0
transformers>=4.42.1
datasets>=2.14.0
numpy>=1.21.0
pandas>=1.3.0
PyYAML>=6.0
tqdm>=4.64.0
jupyter>=1.0.0
notebook>=6.4.0

# Patent data processing dependencies
nltk>=3.8
spacy>=3.6.0
scikit-learn>=1.3.0
lxml>=4.9.0
openpyxl>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Unsloth (optional)
# unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git

# Development tools
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
isort>=5.12.0
pytest-cov>=4.1.0